Update manifests/lens.pp
Set the path of custom lens to "$lens_dest/dist". At least on our ubuntu servers this was needed for augeas to catch it up. Otherwise we are getting 

     Loading failed for one or more files, see debug for /augeas//error output

If you want to specify the custom dir with *load_path* on augeas resource then at least document this somewhere. Note that I couldn't test this on redhat based systems and it might differ there.